+++ examples/evaluation/document-search/basic/evaluate.py
@@ -50,8 +50,8 @@ config = {
                     "type": "ragbits.document_search.ingestion.providers.unstructured:UnstructuredDefaultProvider",
                 },
             },
-            "processing_strategy": {
-                "type": "ragbits.document_search.ingestion.processor_strategies.batched:BatchedAsyncProcessing",
+            "ingest_strategy": {
+                "type": "ragbits.document_search.ingestion.strategies.batched:BatchedIngestStrategy",
                 "config": {
                     "batch_size": 10,
                 },