+++ packages/ragbits-document-search/src/ragbits/document_search/_main.py
@@ -1,6 +1,6 @@
-import warnings
-from collections.abc import Sequence
+from collections.abc import Iterable, Sequence
 from pathlib import Path
+from types import ModuleType
 from typing import Any, ClassVar
 
 from pydantic import BaseModel, Field
@@ -29,11 +29,11 @@ from ragbits.document_search.ingestion import intermediate_handlers
 from ragbits.document_search.ingestion.document_processor import DocumentProcessorRouter
 from ragbits.document_search.ingestion.intermediate_handlers.base import BaseIntermediateHandler
 from ragbits.document_search.ingestion.intermediate_handlers.images import ImageIntermediateHandler
-from ragbits.document_search.ingestion.processor_strategies import (
-    ProcessingExecutionStrategy,
-    SequentialProcessing,
+from ragbits.document_search.ingestion.strategies import (
+    IngestStrategy,
+    SequentialIngestStrategy,
 )
-from ragbits.document_search.ingestion.providers.base import BaseProvider
+from ragbits.document_search.ingestion.strategies.base import IngestExecutionResult
 from ragbits.document_search.retrieval.rephrasers.base import QueryRephraser
 from ragbits.document_search.retrieval.rephrasers.noop import NoopQueryRephraser
 from ragbits.document_search.retrieval.rerankers.base import Reranker, RerankerOptions
@@ -58,7 +58,7 @@ class DocumentSearchConfig(BaseModel):
     vector_store: ObjectContructionConfig
     rephraser: ObjectContructionConfig = ObjectContructionConfig(type="NoopQueryRephraser")
     reranker: ObjectContructionConfig = ObjectContructionConfig(type="NoopReranker")
-    processing_strategy: ObjectContructionConfig = ObjectContructionConfig(type="SequentialProcessing")
+    ingest_strategy: ObjectContructionConfig = ObjectContructionConfig(type="SequentialIngestStrategy")
     providers: dict[str, ObjectContructionConfig] = {}
     intermediate_element_handlers: dict[str, ObjectContructionConfig] = {}
 
@@ -76,33 +76,33 @@ class DocumentSearch(WithConstructionConfig):
         3. Uses Reranker to rerank the chunks.
     """
 
-    # WithConstructionConfig configuration
-    default_module: ClassVar = document_search
-    configuration_key: ClassVar = "document_search"
+    default_module: ClassVar[ModuleType | None] = document_search
+    configuration_key: ClassVar[str] = "document_search"
 
     vector_store: VectorStore
     query_rephraser: QueryRephraser
     reranker: Reranker
-    document_processor_router: DocumentProcessorRouter
-    processing_strategy: ProcessingExecutionStrategy
-    intermediate_element_handlers: dict[type[IntermediateElement], BaseIntermediateHandler]
+
+    ingest_strategy: IngestStrategy
+    parser_router: DocumentProcessorRouter
+    enricher_router: dict[type[IntermediateElement], BaseIntermediateHandler]
 
     def __init__(
         self,
         vector_store: VectorStore,
         query_rephraser: QueryRephraser | None = None,
         reranker: Reranker | None = None,
-        document_processor_router: DocumentProcessorRouter | None = None,
-        processing_strategy: ProcessingExecutionStrategy | None = None,
-        intermediate_element_handlers: dict[type[IntermediateElement], BaseIntermediateHandler] | None = None,
+        ingest_strategy: IngestStrategy | None = None,
+        parser_router: DocumentProcessorRouter | None = None,
+        enricher_router: dict[type[IntermediateElement], BaseIntermediateHandler] | None = None,
     ) -> None:
         self.vector_store = vector_store
         self.query_rephraser = query_rephraser or NoopQueryRephraser()
         self.reranker = reranker or NoopReranker()
-        self.document_processor_router = document_processor_router or DocumentProcessorRouter.from_config()
-        self.processing_strategy = processing_strategy or SequentialProcessing()
-        self.intermediate_element_handlers = intermediate_element_handlers or {
-            IntermediateImageElement: ImageIntermediateHandler(llm=get_preferred_llm(llm_type=LLMType.VISION))
+        self.ingest_strategy = ingest_strategy or SequentialIngestStrategy()
+        self.parser_router = parser_router or DocumentProcessorRouter.from_config()
+        self.enricher_router = enricher_router or {
+            IntermediateImageElement: ImageIntermediateHandler(llm=get_preferred_llm(llm_type=LLMType.VISION)),
         }
 
     @classmethod
@@ -125,23 +125,32 @@ class DocumentSearch(WithConstructionConfig):
         query_rephraser = QueryRephraser.subclass_from_config(model.rephraser)
         reranker: Reranker = Reranker.subclass_from_config(model.reranker)
         vector_store: VectorStore = VectorStore.subclass_from_config(model.vector_store)
-        processing_strategy = ProcessingExecutionStrategy.subclass_from_config(model.processing_strategy)
 
-        providers_config = DocumentProcessorRouter.from_dict_to_providers_config(model.providers)
-        document_processor_router = DocumentProcessorRouter.from_config(providers_config)
-
-        handlers = {
-            import_by_path(element_type, element): import_by_path(
-                handler_config["type"], intermediate_handlers
-            ).from_config(handler_config["config"])
+        ingest_strategy = IngestStrategy.subclass_from_config(model.ingest_strategy)
+        parser_config = DocumentProcessorRouter.from_dict_to_providers_config(model.providers)
+        parser_router = DocumentProcessorRouter.from_config(parser_config)
+        enricher_router = {
+            import_by_path(element_type, element): (
+                import_by_path(handler_config["type"], intermediate_handlers).from_config(handler_config["config"])
+            )
             for element_type, handler_config in config.get("intermediate_handlers", {}).items()
         }
 
-        return cls(vector_store, query_rephraser, reranker, document_processor_router, processing_strategy, handlers)
+        return cls(
+            vector_store=vector_store,
+            query_rephraser=query_rephraser,
+            reranker=reranker,
+            ingest_strategy=ingest_strategy,
+            parser_router=parser_router,
+            enricher_router=enricher_router,
+        )
 
     @classmethod
     def preferred_subclass(
-        cls, config: CoreConfig, factory_path_override: str | None = None, yaml_path_override: Path | None = None
+        cls,
+        config: CoreConfig,
+        factory_path_override: str | None = None,
+        yaml_path_override: Path | None = None,
     ) -> Self:
         """
         Tries to create an instance by looking at project's component prefferences, either from YAML
@@ -214,81 +223,26 @@ class DocumentSearch(WithConstructionConfig):
             )
             return outputs.search_results
 
-    async def ingest(
-        self,
-        documents: str | Sequence[DocumentMeta | Document | Source],
-        document_processor: BaseProvider | None = None,
-    ) -> None:
-        """Ingest documents into the search index.
+    @traceable
+    async def ingest(self, documents: str | Iterable[DocumentMeta | Document | Source]) -> IngestExecutionResult:
+        """
+        Ingest documents into the search index.
 
         Args:
             documents: Either:
-                - A sequence of `Document`, `DocumentMetadata`, or `Source` objects
+                - A iterable of `Document`, `DocumentMetadata`, or `Source` objects
                 - A source-specific URI string (e.g., "gcs://bucket/*") to specify source location(s), for example:
                     - "file:///path/to/files/*.txt"
                     - "gcs://bucket/folder/*"
                     - "huggingface://dataset/split/row"
-            document_processor: The document processor to use. If not provided, the document processor will be
-                determined based on the document metadata.
+
+        Returns:
+            The ingest execution result.
         """
-        with trace(
-            documents=documents,
-            document_processor=document_processor,
-            document_processor_router=self.document_processor_router,
-            processing_strategy=self.processing_strategy,
-        ):
-            if isinstance(documents, str):
-                sources: Sequence[DocumentMeta | Document | Source] = await SourceResolver.resolve(documents)
-            else:
-                sources = documents
-            raw_elements = await self.processing_strategy.process_documents(
-                sources, self.document_processor_router, document_processor
-            )
-
-            intermediate_elements: dict[type, list[IntermediateElement]] = {}
-            for raw_element in raw_elements:
-                if isinstance(raw_element, IntermediateElement):
-                    intermediate_elements.setdefault(type(raw_element), []).append(raw_element)
-
-            elements = [raw_element for raw_element in raw_elements if not isinstance(raw_element, IntermediateElement)]
-
-            for element_type, element_list in intermediate_elements.items():
-                handler = self.intermediate_element_handlers.get(element_type)
-                if handler:
-                    processed_elements = await handler.process(element_list)
-                    elements.extend(processed_elements)
-                else:
-                    warnings.warn(
-                        f"No handler found for {element_type.__name__}. Keeping unprocessed elements.", RuntimeWarning
-                    )
-
-            await self._remove_entries_with_same_sources(elements)
-            await self.insert_elements(elements)
-
-    async def _remove_entries_with_same_sources(self, elements: list[Element]) -> None:
-        """
-        Remove entries from the vector store whose source id is present in the elements' metadata.
-
-        Args:
-            elements: List of elements whose source ids will be checked and removed from the vector store if present.
-        """
-        unique_source_ids = {element.document_meta.source.id for element in elements}
-
-        ids_to_delete = []
-        # TODO: Pass 'where' argument to the list method to filter results and optimize search
-        for entry in await self.vector_store.list():
-            if entry.metadata.get("document_meta", {}).get("source", {}).get("id") in unique_source_ids:
-                ids_to_delete.append(entry.id)
-
-        if ids_to_delete:
-            await self.vector_store.remove(ids_to_delete)
-
-    @traceable
-    async def insert_elements(self, elements: list[Element]) -> None:
-        """
-        Insert Elements into the vector store.
-
-        Args:
-            elements: The list of Elements to insert.
-        """
-        await self.vector_store.store([element.to_vector_db_entry() for element in elements])
+        resolved_documents = await SourceResolver.resolve(documents) if isinstance(documents, str) else documents
+        return await self.ingest_strategy(
+            documents=resolved_documents,
+            vector_store=self.vector_store,
+            parser_router=self.parser_router,
+            enricher_router=self.enricher_router,
+        )