+++ packages/ragbits-core/CHANGELOG.md
@@ -1,10 +1,12 @@
 # CHANGELOG
 
 ## Unreleased
-- improve cli trace handler
-- added traceable to some method
-- Add support for images in few shot prompts (#155)
+
 - BREAKING CHANGE: Vector Stores are now responsible for creating embeddings (#288)
+- Qdrant vector store can now be serialized during Ray processing (#394)
+- Improve cli trace handler
+- Add traceable to some method
+- Add support for images in few shot prompts (#155)
 - Add instruction on how to use local servers for LLMs (#395).
 - Introduce intermediate image elements (#139)
 - Correct typos in doc strings (#398)